requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
ratelimit==2.2.1
SQLAlchemy==2.0.25
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
# LangChain and AI dependencies - updated to compatible versions
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-core==0.1.16
openai>=1.10.0,<2.0.0
# FastAPI for web interface
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
# Additional utilities
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3 